+++ src/plugins/intel_cpu/src/transformations/transformation_pipeline.cpp
@@ -10,6 +10,7 @@
 #include <algorithm>
 #include <cstddef>
 #include <cstdint>
+#include <optional>
 #include <memory>
 #include <set>
 #include <vector>
@@ -134,10 +135,12 @@
 
 // LPT transformations
 #include "low_precision/add.hpp"
+#include "low_precision/avg_pool.hpp"
 #include "low_precision/convert_subtract_constant.hpp"
 #include "low_precision/low_precision.hpp"
 #include "low_precision/multiply_to_group_convolution.hpp"
 #include "low_precision/network_helper.hpp"
+#include "low_precision/resolve_precision_attribute.hpp"
 #include "low_precision/rt_info/bias_attribute.hpp"
 #include "transformations/low_precision/mark_dequantization_subgraph.hpp"
 
@@ -249,6 +252,10 @@
 #    include "transformations/cpu_opset/common/op/sdpa.hpp"
 #    include "transformations/cpu_opset/common/pass/decompose_integer_divide.hpp"
 #    include "transformations/utils.hpp"
+#    if defined(OV_CPU_WITH_ACL)
+#        include "nodes/executors/acl/acl_pooling.hpp"
+#        include "nodes/executors/acl/acl_utils.hpp"
+#    endif
 #else
 #    include "cpu/x64/cpu_isa_traits.hpp"
 #    include "openvino/op/gru_sequence.hpp"
@@ -282,6 +289,123 @@
 namespace ov::intel_cpu {
 
 using const_node_ptr = const std::shared_ptr<const ov::Node>;
+
+#if defined(OPENVINO_ARCH_ARM) || defined(OPENVINO_ARCH_ARM64)
+namespace {
+
+#if defined(OV_CPU_WITH_ACL)
+void copy_pool_attributes(std::vector<ptrdiff_t>& internal_attribute, const std::vector<size_t>& external_attribute) {
+    for (const auto attribute : external_attribute) {
+        internal_attribute.push_back(static_cast<ptrdiff_t>(attribute));
+    }
+}
+
+std::optional<PoolingAttrs> get_avg_pooling_attrs(const std::shared_ptr<const ov::Node>& node) {
+    const auto avg_pool = ov::as_type_ptr<const ov::op::util::AvgPoolBase>(node);
+    if (!avg_pool) {
+        return std::nullopt;
+    }
+
+    PoolingAttrs pooling_attrs;
+    pooling_attrs.algorithm = Algorithm::PoolingAvg;
+    pooling_attrs.exclude_pad = avg_pool->get_exclude_pad();
+    pooling_attrs.rounding = avg_pool->get_rounding_type();
+    copy_pool_attributes(pooling_attrs.kernel, avg_pool->get_kernel());
+    copy_pool_attributes(pooling_attrs.stride, avg_pool->get_strides());
+    if (const auto avg_pool_v16 = ov::as_type_ptr<const ov::op::v16::AvgPool>(node)) {
+        copy_pool_attributes(pooling_attrs.dilation, avg_pool_v16->get_dilations());
+    } else {
+        pooling_attrs.dilation.resize(pooling_attrs.kernel.size(), 1);
+    }
+    copy_pool_attributes(pooling_attrs.data_pad_begin, avg_pool->get_pads_begin());
+    copy_pool_attributes(pooling_attrs.data_pad_end, avg_pool->get_pads_end());
+    pooling_attrs.auto_pad = any_of(avg_pool->get_auto_pad(), ov::op::PadType::SAME_LOWER, ov::op::PadType::SAME_UPPER);
+    return pooling_attrs;
+}
+
+bool is_acl_supported_int8_avg_pool_fq_chain(const std::shared_ptr<const ov::Node>& avg_pool,
+                                             const std::shared_ptr<ov::op::v0::FakeQuantize>& fq,
+                                             const std::vector<ov::element::Type>& defaultPrecisions) {
+    if (!avg_pool || !fq || avg_pool->get_input_partial_shape(0).is_dynamic() || avg_pool->get_output_partial_shape(0).is_dynamic()) {
+        return false;
+    }
+
+    const auto pooling_attrs = get_avg_pooling_attrs(avg_pool);
+    if (!pooling_attrs) {
+        return false;
+    }
+
+    const auto dequantization = ov::pass::low_precision::NetworkHelper::getDequantization(avg_pool, defaultPrecisions);
+    if (dequantization.empty() || !any_of(dequantization.data.get_element_type(), ov::element::Type_t::u8, ov::element::Type_t::i8)) {
+        return false;
+    }
+
+    const auto resolved_precision = ov::pass::low_precision::ResolvePrecisionAttribute::getDataPrecision(fq);
+    if (resolved_precision.empty() || !any_of(resolved_precision.precision, ov::element::Type_t::u8, ov::element::Type_t::i8)) {
+        return false;
+    }
+
+    const auto& input_shape = avg_pool->get_input_shape(0);
+    const auto& output_shape = avg_pool->get_output_shape(0);
+    if (input_shape.size() == 5U) {
+        return false;
+    }
+
+    constexpr auto data_layout = arm_compute::DataLayout::NCHW;
+    arm_compute::TensorInfo src_tensor_info(shapeCast(input_shape),
+                                            1,
+                                            convertToQuantizedType(precisionToAclDataType(dequantization.data.get_element_type())),
+                                            data_layout);
+    arm_compute::TensorInfo dst_tensor_info(shapeCast(output_shape),
+                                            1,
+                                            convertToQuantizedType(precisionToAclDataType(resolved_precision.precision)),
+                                            data_layout);
+
+    arm_compute::PoolingLayerInfo pool_info;
+    return AclPoolingExecutor::isSupported(src_tensor_info,
+                                           dst_tensor_info,
+                                           *pooling_attrs,
+                                           input_shape.size(),
+                                           1,
+                                           data_layout,
+                                           nullptr,
+                                           &pool_info,
+                                           nullptr,
+                                           false);
+}
+#endif
+
+bool should_skip_avg_pool_transformation_on_arm(const const_node_ptr& node,
+                                                [[maybe_unused]] const std::vector<ov::element::Type>& defaultPrecisions) {
+    const auto avg_pool = ov::as_type_ptr<const ov::op::util::AvgPoolBase>(node);
+    if (!avg_pool) {
+        return false;
+    }
+
+#if defined(OV_CPU_WITH_ACL)
+    for (const auto& consumer : avg_pool->output(0).get_target_inputs()) {
+        const auto fq = ov::as_type_ptr<ov::op::v0::FakeQuantize>(consumer.get_node()->shared_from_this());
+        if (!fq) {
+            continue;
+        }
+
+        const auto resolved_precision = ov::pass::low_precision::ResolvePrecisionAttribute::getDataPrecision(fq);
+        if (resolved_precision.empty() ||
+            !any_of(resolved_precision.precision, ov::element::Type_t::u8, ov::element::Type_t::i8)) {
+            continue;
+        }
+
+        if (!is_acl_supported_int8_avg_pool_fq_chain(avg_pool, fq, defaultPrecisions)) {
+            return true;
+        }
+    }
+#endif
+
+    return false;
+}
+
+}  // namespace
+#endif
 
 bool Transformations::is_decompression_multiply(const_node_ptr& node) {
     auto is_1x1_conv = [](const ov::Node* node) {
@@ -980,6 +1104,12 @@ void Transformations::runLptPasses(const std::vector<ov::element::Type>& default
             return ov::marked_as_bias(node);
         },
         AddTransformation);
+    CPU_SET_CALLBACK_ARM(
+        lptManager,
+        [&defaultPrecisions](const_node_ptr& node) -> bool {
+            return should_skip_avg_pool_transformation_on_arm(node, defaultPrecisions);
+        },
+        AvgPoolTransformation);
     CPU_SET_CALLBACK_ARM(
         lptManager,
         [](const_node_ptr& node) -> bool {
